refactor(ui): split state for projectsList and single project
In theory, we could use the same state for both. But the API is not
consistent in what it returns for a resource, depending on the endpoint
it's coming from.

By storing the states separately, we get a little more "Loading" time,
but we don't have to deal with absence of some properties when getting
on the single project page.