fix(ui): issue when deleting attribute from table form
When tracking the items of a list by index, Vue tries to reuse the same
components to render the same index when deleting an item. It's usually
more efficient, but can cause issues with the components' state... which
is what happened here.

I added a virtual unique ID for each attribute so that Vue can track
them based on that.