fix(ui): issue with native form validation pattern on Firefox
It seems like the implementation of pattern validation in Firefox is
less efficient than in Chrome and crashes the app when we abuse it with
a long pattern and a long input.

I switched to a manual validation... which is probably a better solution
in the end. Although it has the downside of not preventing you from
submitting the form.

We should probably look into a proper validation library (e.g.
vee-validate) at some point.